refactor(dart/transform): Remove index_static from hello_world
index_static.js & index_static.html are unnecessary in Js and are now
essentially generated via the Dart transformer. The `angular`
transformer is specified in `examples/pubspec.yaml`; use `pub build` to
create a transformed application that does not use dart:mirrors.

Closes #495
