+++ modules/examples/pubspec.yaml
@@ -16,8 +16,7 @@ dev_dependencies:
     path: ../benchpress
 transformers:
   - angular2:
-      entry_points: web/src/hello_world/index_common.dart
-      reflection_entry_points: web/src/hello_world/index.dart
+      entry_points: web/src/hello_world/index.dart
   - $dart2js:
       minify: false
       #commandLineOptions: [--trust-type-annotations, --trust-primitives, --dump-info]
