+++ modules/examples/src/hello_world/index.html
@@ -1,6 +1,6 @@
 <!doctype html>
 <html>
-  <title>Hello Angular 2.0 (Reflection)</title>
+  <title>Hello Angular 2.0</title>
 <body>
   <hello-app>
     Loading...
