--- modules/examples/src/hello_world/index_static.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html>
-  <title>Hello Angular 2.0 (Static)</title>
-<body>
-  <hello-app>
-    Loading...
-  </hello-app>
-
-  $SCRIPTS$
-</body>
-</html>
